Added godoc commends for embedded methods and types to extend commands. Moved constants to internal directory
