+++ internal/modules/admin/commands.go
@@ -17,8 +17,8 @@ package admin
 import (
 	"errors"
 	"fmt"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
+	"github.com/echovault/echovault/internal/constants"
 	"github.com/gobwas/glob"
 	"slices"
 	"strings"
