+++ internal/modules/connection/commands.go
@@ -17,8 +17,8 @@ package connection
 import (
 	"errors"
 	"fmt"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
+	"github.com/echovault/echovault/internal/constants"
 )
 
 func handlePing(params internal.HandlerFuncParams) ([]byte, error) {
