+++ internal/modules/pubsub/commands.go
@@ -17,8 +17,8 @@ package pubsub
 import (
 	"errors"
 	"fmt"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
+	"github.com/echovault/echovault/internal/constants"
 	"strings"
 )
