+++ internal/modules/sorted_set/commands.go
@@ -18,8 +18,8 @@ import (
 	"cmp"
 	"errors"
 	"fmt"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
+	"github.com/echovault/echovault/internal/constants"
 	"math"
 	"slices"
 	"strconv"
