+++ internal/modules/string/commands.go
@@ -17,8 +17,8 @@ package str
 import (
 	"errors"
 	"fmt"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
+	"github.com/echovault/echovault/internal/constants"
 )
 
 func handleSetRange(params internal.HandlerFuncParams) ([]byte, error) {
