+++ internal/modules/string/key_funcs.go
@@ -16,8 +16,8 @@ package str
 
 import (
 	"errors"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
+	"github.com/echovault/echovault/internal/constants"
 )
 
 func setRangeKeyFunc(cmd []string) (internal.KeyExtractionFuncResult, error) {
