+++ test/modules/generic/commands_test.go
@@ -19,11 +19,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/echovault"
 	"github.com/echovault/echovault/internal"
 	"github.com/echovault/echovault/internal/clock"
 	"github.com/echovault/echovault/internal/config"
+	"github.com/echovault/echovault/internal/constants"
 	"github.com/tidwall/resp"
 	"net"
 	"reflect"
