+++ test/modules/hash/commands_test.go
@@ -19,10 +19,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/echovault"
 	"github.com/echovault/echovault/internal"
 	"github.com/echovault/echovault/internal/config"
+	"github.com/echovault/echovault/internal/constants"
 	"github.com/tidwall/resp"
 	"net"
 	"reflect"
